Basic support for Cabal
When given a Cabal project directory (and, optionally, a build target), we load
the cabal configuration and initialize the session using the modules, default
extensions, GHC options, and package dependencies and package DB (unless
overwritten on the command line).

Some caveats:

* Since ide-backend handles it own shadow copy of a project's source, this
  means that the list of modules listed in other-modules (together with
  exposed-modules) must be complete, or else GHC won't be able to find all
  modules.
* The package DB specified in a Cabal file is relative ("global"/"user") and
  therefore may not be the one that is actually required to load a project
  (if the default package DB of the compiler used to compile the ide-backend
  server does not match the package DB of the compiler used to configure
  the project). For now this can be overridden on the command line by specifying
  a package-db, but we might have to be smarter about this during session
  initialization.

The session gets initialized with the modules in the Cabal file, but they are
not compiled until the first call to updateSession; this gives the editor a
change to deal to set things up to deal with progress messages nicely; if the
editor wants to load all modules immediately, it can send

    { "request": "updateSession", "update": [] }

(i.e., an empty update) to bring the session up to date.