Adds --ide-backend-tools-path and --extra-path
With `--ide-backend-tools-path` one can indicate where to search for
`ide-backend-server` and `ide-backend-exe-cabal` binaries. If not given,
defaults to system path (but note that the system path won't be used
as well if a value for this arg is given).

With `--extra-path` one can provide an additional path that will be
additionally used when searching for `ghc`, `ide-backend-server`,
etc. It replaces the previous `--path` argument.