Optional seq fld in requests, quoted by responses
Implements the proposal in (#39), namely, every request accepts
an additional field `seq`, whose value can be any JSON value and
if present, it is included in a field `seq` of every response associated
to the request. Clients that include no `seq` field in their requests will
get no `seq` field in their responses.

This increments the protocol version to 0.1.0.