Make it possible to jump back from M-.
SLIME uses etags' tag-marker ring in the function which is called from
its version of M-. in order to use etags' facilities for popping back.
This does the same.

Unlike SLIME, it does not bind M-, to (a function which calls)
pop-tag-mark.  It's already bound by default to M-* and while I'll
probably rebind it myself locally I don't really see a good reason to do
so by default.